Task 3: Fix conversion warnings from strict flags
Systematically fix conversion and shadow warnings surfaced by -Wconversion and -Werror:

Vendor header pragmas (no code changes, silence warnings at include site):
- Wrap stb_image.h include with -Wconversion pragma (src/image/decode.cpp)
- Wrap stb_truetype/resize2/write includes with pragma (src/gfx/text.cpp, src/image/thumbnail.cpp)
- Wrap libavcodec/libavformat/libavutil includes with pragma in headers and source files
  (src/media/audio_decoder.*, video_decoder.*, chunk_avio.*, mem_avio.*, audio_interleave.*)
- Wrap test FFmpeg includes (tests/media/test_av_link.cpp)

Project code conversion fixes (semantics-preserving):
- byte_io.h:get() — explicit cast of shift result to avoid lossy intermediate
- advanced_search_screen.cpp:draw_dropdown() — use float literals (6.0f, 12.0f) and cast int to float
- search_model.cpp:ascii_lower() — explicit cast char arithmetic result back to char
- test_text.cpp — cast size_t to int for comparison with vector size

All 618 tests pass; ASan and Release builds green.

+++ src/ui/advanced_search_screen.cpp
@@ -101,7 +101,7 @@ void draw_dropdown(gfx::Renderer& r, gfx::FontAtlas& font, const std::vector<std
     // overlaps the labels beneath. Each suggestion's ink is vertically centred in
     // its slot via text_top_for_center (draw_text's y is the glyph-cell top, so the
     // ink would otherwise sit low and poke out of the panel).
-    const SDL_FRect panel{x - 6, y, colw + 12, n * ROW};
+    const SDL_FRect panel{x - 6.0f, y, colw + 12.0f, static_cast<float>(n) * ROW};
     r.draw_round_rect(panel, RADIUS_SMALL, SURFACE);
     r.draw_round_rect(panel, RADIUS_SMALL, ACCENT, /*filled*/ false);
     for (int i = 0; i < n; ++i) {
